Create MapExample.c
Because I don't have a whole lot of time on my hands and I don't want to hold up the project anymore than I already have, this is my template for creating the map we need.  The section that creates the map starts on line 75 and ends on line 135.  This is in C, so less lines should be needed to do the same thing in Java.  I'll try to work on it more tomorrow (technically today), but I'm not going to make any promises.  I figured I'd post this for anyone else who wants to give this a shot.