Code now prints the linked list
I'm thinking we are only going to need to code the toString() function in the NodeList.java file.  This currently prints out the map from head to tail, and that's basically all we are doing with this whole project--only we're printing out the map in different ways and counting as we go along.